+++ R/wconf-package.R
@@ -6,6 +6,9 @@
 #'   ConfusionMatrix object or a simple matrix, according to one of several
 #'   weighting schemas and optionally prints the weighted accuracy score.
 #'
+#'   balancedaccuracy - calculates classification accuracy scores for imbalanced
+#'   data using sine-based formulas proposed by Starovoitov and Golub (2020).
+#'
 #' @examples
 #' # Generate sample data
 #' m = matrix(c(70,0,0,10,10,0,5,3,2), ncol = 3, nrow=3)
@@ -15,13 +18,19 @@
 #' weightmatrix(n=4, weight.type="geometric", geometric.multiplier = 2)
 #' # Compute the weighted confusion matrix
 #' wconfusionmatrix(m, weight.type = "arithmetic", print.weighted.accuracy = TRUE)
+#' # Calculate accuracy metrics for imbalanced data
+#' balancedaccuracy(m, print.scores = TRUE)
 #'
 #' @author Alexandru Monahov
+#'
 #' @references Kuhn, M. (2008). Building Perspective Models in R Using the caret
 #' Package. Journal of Statistical Software, 28(5), 1-26.
 #'
 #' Monahov, A. (2021). Model Evaluation with Weighted Threshold Optimization
 #' (and the "mewto" R package), Machine Learning eJournal, SSRN.
+#'
+#' Starovoitov, V., Golub, Y. (2020). New Function for Estimating Imbalanced
+#' Data Classification Results. Pattern Recognition and Image Analysis, 295–302.
 #'
 #' Van de Velden, M. et al. (2023). A general framework for implementing
 #' distances for categorical variables. arXiv.